pyproject.toml
[tool.poetry]
name = "f28379d-drv8305-dual-comm"
version = "0.1.0"
description = ""
authors = ["wenliangyeoh <wlyeoh@cc.saga-u.ac.jp>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"
crc = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
pytest = "^7.1.3"
Sphinx = "^5.2.3"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
